Backported CORDA-1459 to the release-v3
