+++ finance/src/main/kotlin/net/corda/finance/contracts/FinanceTypes.kt
@@ -16,7 +16,6 @@ import net.corda.core.flows.FlowException
 import net.corda.core.identity.Party
 import net.corda.core.serialization.CordaSerializable
 import net.corda.core.transactions.TransactionBuilder
-import net.corda.finance.contracts.asset.CommodityContract
 import java.math.BigDecimal
 import java.time.DayOfWeek
 import java.time.LocalDate
